Enum clang::diagnostic::FixIt

pub enum FixIt<'tu> {
    Deletion(SourceRange<'tu>),
    Insertion(SourceLocation<'tu>, String),
    Replacement(SourceRange<'tu>, String),
}

A suggested fix for an issue with a source file.

Variants
Deletion(SourceRange<'tu>)

Delete a segment of the source file.

Insertion(SourceLocation<'tu>, String)

Insert a string into the source file.

Replacement(SourceRange<'tu>, String)

Replace a segment of the source file with a string.