Struct clang::diagnostic::Diagnostic

pub struct Diagnostic<'tu> { /* fields omitted */ }

A message from the compiler about an issue with a source file.

Methods

impl<'tu> Diagnostic<'tu>

pub fn get_severity(&self) -> Severity

Returns the severity of this diagnostic.

pub fn get_text(&self) -> String

Returns the text of this diagnostic.

pub fn get_location(&self) -> SourceLocation<'tu>

Returns the source location of this diagnostic.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn get_ranges(&self) -> Vec<SourceRange<'tu>>

Returns the source ranges of this diagnostic.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn get_fix_its(&self) -> Vec<FixIt<'tu>>

Returns the fix-its for this diagnostic.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn get_children(&self) -> Vec<Diagnostic>

Returns the child diagnostics of this diagnostic.

pub fn formatter(&self) -> DiagnosticFormatter<'tu>

Returns a diagnostic formatter that builds a formatted string from this diagnostic.