Struct clang::diagnostic::DiagnosticFormatter

pub struct DiagnosticFormatter<'tu> { /* fields omitted */ }

Builds formatted strings from diagnostics.

Methods

impl<'tu> DiagnosticFormatter<'tu>

fn source_location(&mut self,
source_location: bool)
-> &mut DiagnosticFormatter<'tu>

Sets whether the diagnostic text will be prefixed by the file and line of the source location the diagnostic indicates. This prefix may also contain column and/or source range information.

fn column(&mut self, column: bool) -> &mut DiagnosticFormatter<'tu>

Sets whether the column will be included in the source location prefix.

fn source_ranges(&mut self,
source_ranges: bool)
-> &mut DiagnosticFormatter<'tu>

Sets whether the source ranges will be included to the source location prefix.

fn option(&mut self, option: bool) -> &mut DiagnosticFormatter<'tu>

Sets whether the option associated with the diagnostic (e.g., -Wconversion) will be placed in brackets after the diagnostic text if there is such an option.

fn category_id(&mut self, category_id: bool) -> &mut DiagnosticFormatter<'tu>

Sets whether the category number associated with the diagnostic will be placed in brackets after the diagnostic text if there is such a category number.

fn category_name(&mut self,
category_name: bool)
-> &mut DiagnosticFormatter<'tu>

Sets whether the category name associated with the diagnostic will be placed in brackets after the diagnostic text if there is such a category name.

impl<'tu> DiagnosticFormatter<'tu>

fn format(&self) -> String

Returns a formatted string.