Module clang::documentation

Comments.

Structs

BlockCommand
A block command with zero or more arguments and a paragraph as an argument.

Comment
A comment attached to a declaration.

HtmlStartTag
An HTML start tag.

InlineCommand
An inline command with word-like arguments.

ParamCommand
A \param command.

TParamCommand
A \tparam command.

Enums

CommentChild
A child component of a comment.

InlineCommandStyle
Indicates the appropriate rendering style for an inline command argument.

ParameterDirection
Indicates the parameter passing direction for a \param command.