Enum clang::documentation::CommentChild

pub enum CommentChild {
    BlockCommand(BlockCommand),
    HtmlStartTag(HtmlStartTag),
    HtmlEndTag(String),
    InlineCommand(InlineCommand),
    Paragraph(Vec<CommentChild>),
    ParamCommand(ParamCommand),
    TParamCommand(TParamCommand),
    Text(String),
    VerbatimCommand(Vec<String>),
    VerbatimLineCommand(String),
}

A child component of a comment.

Variants
BlockCommand(BlockCommand)

A block command with zero or more arguments and a paragraph as an argument.

HtmlStartTag(HtmlStartTag)

An HTML start tag.

HtmlEndTag(String)

An HTML end tag.

InlineCommand(InlineCommand)

An inline command with word-like arguments.

Paragraph(Vec<CommentChild>)

A paragraph containing inline content.

ParamCommand(ParamCommand)

A \param command.

TParamCommand(TParamCommand)

A \tparam command.

Text(String)

Plain text.

VerbatimCommand(Vec<String>)

A verbatim command with a closing command.

VerbatimLineCommand(String)

A verbatim command with a single line and no closing command.