Enum clang::CompletionChunk

pub enum CompletionChunk<'r> {
    Optional(CompletionString<'r>),
    CurrentParameter(String),
    Informative(String),
    Placeholder(String),
    ResultType(String),
    Text(String),
    TypedText(String),
    Colon(String),
    Comma(String),
    Equals(String),
    Semicolon(String),
    LeftAngleBracket(String),
    RightAngleBracket(String),
    LeftBrace(String),
    RightBrace(String),
    LeftParenthesis(String),
    RightParenthesis(String),
    LeftSquareBracket(String),
    RightSquareBracket(String),
    HorizontalSpace(String),
    VerticalSpace(String),
}

A piece of a code completion string.

Variants

Optional

An optional piece that could be part of the template but is not required.

CurrentParameter

Text that describes the current parameter when code completion was run on a function call, message send, or template specialization.

Informative

Informative text that should be displayed but not inserted as part of the template.

Placeholder

Text that should be replaced by the user.

ResultType

Text that specifies the result type of the containing result.

Text

Text that should be inserted.

TypedText

Text that the user would be expected to type to get the containing code completion result.

Colon

A colon (':').

Comma

A comma (',').

Equals

An equals sign ('=').

Semicolon

A semicolon (';').

LeftAngleBracket

A left angle bracket ('<').

RightAngleBracket

A right angle bracket ('>').

LeftBrace

A left brace ('{').

RightBrace

A right brace ('}').

LeftParenthesis

A left parentesis ('(')).

RightParenthesis

A right parenthesis (')').

LeftSquareBracket

A left square bracket ('[').

RightSquareBracket

A right square bracket (']').

HorizontalSpace

Horizontal space (e.g., ' ').

VerticalSpace

Vertical space (e.g., '\n').

Methods

impl<'r> CompletionChunk<'r>

fn get_text(&self) -> String

Returns the text associated with this completion chunk.

Panics

this completion chunk is a CompletionChunk::Optional