Enum clang::EvaluationResult

pub enum EvaluationResult {
    Unexposed,
    SignedInteger(i64),
    UnsignedInteger(u64),
    Float(f64),
    String(CString),
    ObjCString(CString),
    CFString(CString),
    Other(CString),
}

The result of evaluating an expression.

Variants
Unexposed

An evaluation result whose specific type is not exposed via this interface.

SignedInteger(i64)

A signed integer evaluation result.

UnsignedInteger(u64)

An unsigned integer evaluation result.

Only produced by libclang 4.0 and later. Earlier versions will always return SignedInteger for integers.

Float(f64)

A floating point number evaluation result.

String(CString)

A string literal evaluation result.

ObjCString(CString)

An Objective-C string literal evaluation result.

CFString(CString)

An Objective-C CFString evaluation result.

Other(CString)

Any other evaluation result whose value can be represented by a string.