Module clang::sonar

Finding C declarations.

Structs

Declaration
A C declaration.

Definition
A C preprocessor definition.

Definitions
An iterator over preprocessor definition declarations.

Enums
An iterator over enum declarations.

Functions
An iterator over function declarations.

Structs
An iterator over struct declarations.

Typedefs
An iterator over typedef declarations.

Unions
An iterator over struct declarations.

Enums

DefinitionValue
The value of a C preprocessor definition.

Functions

find_definitions
Returns an iterator over the simple preprocessor definitions in the supplied entities.

find_enums
Returns an iterator over the enums in the supplied entities.

find_functions
Returns an iterator over the functions in the supplied entities.

find_structs
Returns an iterator over the structs in the supplied entities.

find_typedefs
Returns an iterator over the typedefs in the supplied entities.

find_unions
Returns an iterator over the unions in the supplied entities.