Function clang::sonar::find_structs

Important traits for Structs<'tu>
impl<'tu> Iterator for Structs<'tu> type Item = Declaration<'tu>;

pub fn find_structs<'tu, E: Into<Vec<Entity<'tu>>>>(entities: E) -> Structs<'tu>

Returns an iterator over the structs in the supplied entities.

If a struct is encountered multiple times, only the first instance is included.