Struct clang::source::File

pub struct File<'tu> { /* fields omitted */ }

A source file.

Methods

impl<'tu> File<'tu>

fn get_path(&self) -> PathBuf

Returns the absolute path to this file.

fn get_time(&self) -> time_t

Returns the last modification time for this file.

fn get_id(&self) -> (u64, u64, u64)

Returns a unique identifier for this file.

fn get_module(&self) -> Option<Module<'tu>>

Returns the module containing this file, if any.

fn get_skipped_ranges(&self) -> Vec<SourceRange<'tu>>

Returns the source ranges in this file that were skipped by the preprocessor.

This will always return an empty Vec if the translation unit that contains this file was not constructed with a detailed preprocessing record.

fn is_include_guarded(&self) -> bool

Returns whether this file is guarded against multiple inclusions.

fn get_location(&self, line: u32, column: u32) -> SourceLocation<'tu>

Returns the source location at the supplied line and column in this file.

Panics

line or column is 0

fn get_offset_location(&self, offset: u32) -> SourceLocation<'tu>

Returns the source location at the supplied character offset in this file.

fn get_includes(&self) -> Vec<Entity<'tu>>

Returns the inclusion directives in this file.

fn get_references(&self, entity: Entity<'tu>) -> Vec<Entity<'tu>>

Returns the references to the supplied entity in this file.

fn visit_includes<F: FnMut(Entity<'tu>, SourceRange<'tu>) -> bool>(&self,
f: F)
-> bool

Visits the inclusion directives in this file and returns whether visitation was ended by the callback returning false.

fn visit_references<F: FnMut(Entity<'tu>, SourceRange<'tu>) -> bool>(&self,
entity: Entity<'tu>,
f: F)
-> bool

Visits the references to the supplied entity in this file and returns whether visitation was ended by the callback returning false.