Struct clang::source::SourceLocation

pub struct SourceLocation<'tu> { /* fields omitted */ }

A location in a source file.

Methods

impl<'tu> SourceLocation<'tu>

fn get_expansion_location(&self) -> Location<'tu>

Returns the file, line, column and character offset of this source location.

If this source location is inside a macro expansion, the location of the macro expansion is returned instead.

fn get_file_location(&self) -> Location<'tu>

Returns the file, line, column and character offset of this source location.

If this source location is inside a macro expansion, the location of the macro expansion is returned instead unless this source location is inside a macro argument. In that case, the location of the macro argument is returned.

fn get_presumed_location(&self) -> (String, u32, u32)

Returns the file path, line, and column of this source location taking line directives into account.

fn get_spelling_location(&self) -> Location<'tu>

Returns the file, line, column and character offset of this source location.

fn get_entity(&self) -> Option<Entity<'tu>>

Returns the AST entity at this source location, if any.

fn is_in_main_file(&self) -> bool

Returns whether this source location is in the main file of its translation unit.

fn is_in_system_header(&self) -> bool

Returns whether this source location is in a system header.