Struct clang::CompletionContext

pub struct CompletionContext {
    pub all_types: bool,
    pub all_values: bool,
    pub class_type_values: bool,
    pub dot_members: bool,
    pub arrow_members: bool,
    pub enum_tags: bool,
    pub union_tags: bool,
    pub struct_tags: bool,
    pub class_names: bool,
    pub namespaces: bool,
    pub nested_name_specifiers: bool,
    pub macro_names: bool,
    pub natural_language: bool,
    pub objc_object_values: bool,
    pub objc_selector_values: bool,
    pub objc_property_members: bool,
    pub objc_interfaces: bool,
    pub objc_protocols: bool,
    pub objc_categories: bool,
    pub objc_instance_messages: bool,
    pub objc_class_messages: bool,
    pub objc_selector_names: bool,
}

Indicates which types of results were included in a set of code completion results.

Fields

all_types

Indicates whether all possible types were included.

all_values

Indicates whether all possible values were included.

class_type_values

Indicates whether values that resolve to C++ class types were included.

dot_members

Indicates whether the members of a record that are accessed with the dot operator were included.

arrow_members

Indicates whether the members of a record that are accessed with the arrow operator were included.

enum_tags

Indicates whether enum tags were included.

union_tags

Indicates whether union tags were included.

struct_tags

Indicates whether struct tags were included.

class_names

Indicates whether C++ class names were included.

namespaces

Indicates whether C++ namespaces and namespace aliases were included.

nested_name_specifiers

Indicates whether C++ nested name specifiers were included.

macro_names

Indicates whether macro names were included.

natural_language

Indicates whether natural language results were included.

objc_object_values

Indicates whether values that resolve to Objective-C objects were included.

objc_selector_values

Indicates whether values that resolve to Objective-C selectors were included.

objc_property_members

Indicates whether the properties of an Objective-C object that are accessed with the dot operator were included.

objc_interfaces

Indicates whether Objective-C interfaces were included.

objc_protocols

Indicates whether Objective-C protocols were included.

objc_categories

Indicates whether Objective-C categories were included.

objc_instance_messages

Indicates whether Objective-C instance messages were included.

objc_class_messages

Indicates whether Objective-C class messages were included.

objc_selector_names

Indicates whether Objective-C selector names were included.