Struct clang::FormatOptions

pub struct FormatOptions {
    pub source_location: bool,
    pub column: bool,
    pub source_ranges: bool,
    pub option: bool,
    pub category_id: bool,
    pub category_name: bool,
}

A set of options that determines how a diagnostic is formatted.

Fields

source_location

Indicates whether the diagnostic text will be prefixed by the file and line of the source location the diagnostic indicates. This prefix may also contain column and/or source range information.

column

Indicates whether the column will be included in the source location prefix.

source_ranges

Indicates whether the source ranges will be included to the source location prefix.

option

Indicates whether the option associated with the diagnostic (e.g., -Wconversion) will be placed in brackets after the diagnostic text if there is such an option.

category_id

Indicates whether the category number associated with the diagnostic will be placed in brackets after the diagnostic text if there is such a category number.

category_name

Indicates whether the category name associated with the diagnostic will be placed in brackets after the diagnostic text if there is such a category name.