Struct clang::Index

pub struct Index<'c> { /* fields omitted */ }

A collection of translation units.

Methods

impl<'c> Index<'c>

pub fn new(_: &'c Clang, exclude: bool, diagnostics: bool) -> Index<'c>

Constructs a new Index.

exclude determines whether declarations from precompiled headers are excluded and diagnostics determines whether diagnostics are printed while parsing source files.

pub fn parser<F: Into<PathBuf>>(&'c self, f: F) -> Parser<'c>

Returns a parser for the supplied file.

pub fn set_invocation_emission_path<P: AsRef<Path>>(&'c self, path: P)

Sets the invocation emission path for this index.

pub fn get_thread_options(&self) -> ThreadOptions

Returns the thread options for this index.

pub fn set_thread_options(&mut self, options: ThreadOptions)

Sets the thread options for this index.