Struct clang::Parser

pub struct Parser<'tu> { /* fields omitted */ }

Parses translation units.

Methods

impl<'tu> Parser<'tu>

pub fn cache_completion_results(
&mut self,
cache_completion_results: bool
) -> &mut Parser<'tu>

Sets whether certain code completion results will be cached when the translation unit is reparsed.

This option increases the time it takes to reparse the translation unit but improves code completion performance.

pub fn detailed_preprocessing_record(
&mut self,
detailed_preprocessing_record: bool
) -> &mut Parser<'tu>

Sets whether a detailed preprocessing record will be constructed which tracks all macro definitions and instantiations.

pub fn briefs_in_completion_results(
&mut self,
briefs_in_completion_results: bool
) -> &mut Parser<'tu>

Sets whether documentation comment briefs will be included in code completion results.

pub fn incomplete(&mut self, incomplete: bool) -> &mut Parser<'tu>

Sets whether the translation unit will be considered incomplete.

This option suppresses certain semantic analyses and is typically used when parsing headers with the intent of creating a precompiled header.

pub fn skip_function_bodies(
&mut self,
skip_function_bodies: bool
) -> &mut Parser<'tu>

Sets whether function and method bodies will be skipped.

pub fn keep_going(&mut self, keep_going: bool) -> &mut Parser<'tu>

Sets whether processing will continue after a fatal error is encountered.

pub fn single_file_parse(&mut self, single_file_parse: bool) -> &mut Parser<'tu>

Sets whether incremental processing will be used.

impl<'tu> Parser<'tu>

pub fn arguments<S: AsRef<str>>(&mut self, arguments: &[S]) -> &mut Parser<'tu>

Sets the compiler arguments to provide to libclang.

Any compiler argument that could be supplied to clang may be supplied to this function. However, the following arguments are ignored:

-c

-emit-ast

-fsyntax-only

-o and the following <output>

pub fn unsaved(&mut self, unsaved: &[Unsaved]) -> &mut Parser<'tu>

Sets the unsaved files to use.

pub fn parse(&self) -> Result<TranslationUnit<'tu>, SourceError>

Parses a translation unit.

Failures

an error occurs while deserializing an AST file

libclang crashes

an unknown error occurs