Struct clang::TranslationUnit

pub struct TranslationUnit<'i> { /* fields omitted */ }

A preprocessed and parsed source file.

Methods

impl<'i> TranslationUnit<'i>

pub fn from_ast<F: AsRef<Path>>(
index: &'i Index,
file: F
) -> Result<TranslationUnit<'i>, ()>

Constructs a new TranslationUnit from an AST file.

Failures

an unknown error occurs

pub fn get_diagnostics(&'i self) -> Vec<Diagnostic<'i>>

Returns the diagnostics for this translation unit.

pub fn get_entity(&'i self) -> Entity<'i>

Returns the entity for this translation unit.

pub fn get_file<F: AsRef<Path>>(&'i self, file: F) -> Option<File<'i>>

Returns the file at the supplied path in this translation unit, if any.

pub fn get_memory_usage(&self) -> HashMap<MemoryUsage, usize>

Returns the memory usage of this translation unit.

pub fn get_skipped_ranges(&'i self) -> Vec<SourceRange<'i>>

Returns the source ranges in this translation unit that were skipped by the preprocessor.

This will always return an empty Vec if the translation unit was not constructed with a detailed preprocessing record.

pub fn get_target(&self) -> Target

Returns information about the target for this translation unit.

pub fn annotate(&'i self, tokens: &[Token<'i>]) -> Vec<Option<Entity<'i>>>

Returns the AST entities which correspond to the supplied tokens, if any.

pub fn completer<F: Into<PathBuf>>(
&self,
file: F,
line: u32,
column: u32
) -> Completer

Returns a completer which runs code completion.

pub fn save<F: AsRef<Path>>(&self, file: F) -> Result<(), SaveError>

Saves this translation unit to an AST file.

Failures

errors in the translation unit prevent saving

an unknown error occurs

pub fn reparse(
self,
unsaved: &[Unsaved]
) -> Result<TranslationUnit<'i>, SourceError>

Consumes this translation unit and reparses the source file it was created from with the same compiler arguments that were used originally.

Failures

an error occurs while deserializing an AST file

libclang crashes

an unknown error occurs