Crate clang

Bindings and idiomatic wrapper for libclang.

Version Specific Documentation

3.5.x - Documentation

3.6.x - Documentation

3.7.x - Documentation

The documentation for the non-version specific API is here.

Modules

ffi
The raw bindings to libclang.

Structs

Clang
An empty type which prevents the use of this library from multiple threads simultaneously.

CompilationDatabase
The information used to compile the source files in a project.

CompileCommand
The information used to compile a source file in a project.

CompletionContext
Indicates which types of results were included in a set of code completion results.

CompletionOptions
A set of options that determines how code completion is run.

CompletionResult
A code completion result.

CompletionResults
A set of code completion results.

CompletionString
A semantic string that describes a code completion result.

Diagnostic
A message from the compiler about an issue with a source file.

Entity
An AST entity.

File
A source file.

FormatOptions
A set of options that determines how a diagnostic is formatted.

Index
A collection of translation units.

Location
The file, line, column, and character offset of a source location.

Module
A collection of headers.

ObjCAttributes
Indicates which attributes were applied to an Objective-C property.

ObjCQualifiers
Indicates which qualifiers were applied to an Objective-C method return or parameter type.

ParseOptions
A set of options that determines how a source file is parsed into a translation unit.

PlatformAvailability
The availability of an AST entity on a particular platform.

SourceLocation
A location in a source file.

SourceRange
A half-open range in a source file.

ThreadOptions
A set of options that determines which types of threads should use background priority.

Token
A lexed piece of a source file.

TranslationUnit
A preprocessed and parsed source file.

Type
The type of an AST entity.

Unsaved
The path to and unsaved contents of a previously existing file.

Usr
A Unified Symbol Resolution (USR).

Version
A version number in the form x.y.z.

Enums

Accessibility
Indicates the accessibility of a declaration or base class specifier.

AlignofError
Indicates the error that prevented determining the alignment of a type.

Availability
Indicates the availability of an AST entity.

CallingConvention
Indicates the calling convention specified for a function type.

CompletionChunk
A piece of a code completion string.

EntityKind
Indicates the categorization of an AST entity.

EntityVisitResult
Indicates how a entity visitation should proceed.

FixIt
A suggested fix for an issue with a source file.

Language
Indicates the language used by a declaration.

Linkage
Indicates the linkage of an AST entity.

MemoryUsage
Indicates the usage category of a quantity of memory.

OffsetofError
Indicates the error that prevented determining the offset of a field in a record type.

RefQualifier
Indicates the ref qualifier of a C++ function or method type.

SaveError
Indicates the type of error that prevented the saving of a translation unit to an AST file.

Severity
Indicates the severity of a diagnostic.

SizeofError
Indicates the error that prevented determining the size of a type.

SourceError
Indicates the type of error that prevented the loading of a translation unit from a source file.

StorageClass
Indicates the storage class of a declaration.

TemplateArgument
An argument to a template function specialization.

TokenKind
Indicates the categorization of a token.

TypeKind
Indicates the categorization of a type.

Traits

Nullable
A type which may be null or otherwise invalid.

Implementations