Crate clang

A somewhat idiomatic Rust wrapper for libclang.

Supported Versions

3.5 - Documentation

3.6 - Documentation

3.7 - Documentation

3.8 - Documentation

3.9 - Documentation

Modules

completion
Code completion.

diagnostic
Issues with source files.

documentation
Comments.

sonar
Finding C declarations.

source
Source files, locations, and ranges.

token
Lexed pieces of source files.

Structs

Clang
An empty type which prevents the use of this library from multiple threads simultaneously.

Entity
An AST entity.

Index
A collection of translation units.

ObjCAttributes
Indicates which attributes were applied to an Objective-C property.

ObjCQualifiers
Indicates which qualifiers were applied to an Objective-C method return or parameter type.

Parser
Parses translation units.

PlatformAvailability
The availability of an AST entity on a particular platform.

ThreadOptions
A set of options that determines which types of threads should use background priority.

TranslationUnit
A preprocessed and parsed source file.

Type
The type of an AST entity.

Unsaved
The path to and unsaved contents of a previously existing file.

Usr
A Unified Symbol Resolution (USR).

Version
A version number in the form x.y.z.

Enums

Accessibility
Indicates the accessibility of a declaration or base class specifier.

AlignofError
Indicates the error that prevented determining the alignment of a type.

Availability
Indicates the availability of an AST entity.

CallingConvention
Indicates the calling convention specified for a function type.

EntityKind
Indicates the categorization of an AST entity.

EntityVisitResult
Indicates how a entity visitation should proceed.

Language
Indicates the language used by a declaration.

Linkage
Indicates the linkage of an AST entity.

MemoryUsage
Indicates the usage category of a quantity of memory.

OffsetofError
Indicates the error that prevented determining the offset of a field in a record type.

RefQualifier
Indicates the ref qualifier of a C++ function or method type.

SaveError
Indicates the type of error that prevented the saving of a translation unit to an AST file.

SizeofError
Indicates the error that prevented determining the size of a type.

SourceError
Indicates the type of error that prevented the loading of a translation unit from a source file.

TypeKind
Indicates the categorization of a type.

Functions

get_version
Returns the version string for the version of libclang in use.