Enum clang::completion::CompletionChunk

pub enum CompletionChunk<'r> {
    Colon,
    Comma,
    Equals,
    Semicolon,
    LeftAngleBracket,
    RightAngleBracket,
    LeftBrace,
    RightBrace,
    LeftParenthesis,
    RightParenthesis,
    LeftSquareBracket,
    RightSquareBracket,
    HorizontalSpace(String),
    VerticalSpace(String),
    CurrentParameter(String),
    Informative(String),
    Placeholder(String),
    ResultType(String),
    Text(String),
    TypedText(String),
    Optional(CompletionString<'r>),
}

A piece of a code completion string.

Variants
Colon

A colon (':').

Comma

A comma (',').

Equals

An equals sign ('=').

Semicolon

A semicolon (';').

LeftAngleBracket

A left angle bracket ('<').

RightAngleBracket

A right angle bracket ('>').

LeftBrace

A left brace ('{').

RightBrace

A right brace ('}').

LeftParenthesis

A left parentesis ('(')).

RightParenthesis

A right parenthesis (')').

LeftSquareBracket

A left square bracket ('[').

RightSquareBracket

A right square bracket (']').

HorizontalSpace(String)

Horizontal space (e.g., ' ').

VerticalSpace(String)

Vertical space (e.g., '\n').

CurrentParameter(String)

Text that describes the current parameter when code completion was run on a function call, message send, or template specialization.

Informative(String)

Informative text that should be displayed but not inserted as part of the template.

Placeholder(String)

Text that should be replaced by the user.

ResultType(String)

Text that specifies the result type of the containing result.

Text(String)

Text that should be inserted.

TypedText(String)

Text that the user would be expected to type to get the containing code completion result.

Optional(CompletionString<'r>)

An optional piece that could be part of the template but is not required.

Methods

impl<'r> CompletionChunk<'r>

pub fn get_text(&self) -> Option<String>

Returns the text associated with this completion chunk if this chunk is not optional.

pub fn is_optional(&self) -> bool

Returns whether this completion chunk is optional.