Struct clang::completion::CompletionResults

pub struct CompletionResults { /* fields omitted */ }

A set of code completion results.

Methods

impl CompletionResults

fn get_diagnostics<'tu>(&self,
tu: &'tu TranslationUnit<'tu>)
-> Vec<Diagnostic<'tu>>

Returns the diagnostics that were produced prior to the code completion context for this set of code completion results.

fn get_context(&self) -> Option<CompletionContext>

Returns the code completion context for this set of code completion results, if any.

fn get_container_kind(&self) -> Option<(EntityKind, bool)>

Returns the categorization of the entity that contains the code completion context for this set of code completion results and whether that entity is incomplete, if applicable.

fn get_objc_selector(&self) -> Option<String>

Returns the selector or partial selector that has been entered this far for the Objective-C message send context for this set of code completion results.

fn get_usr(&self) -> Option<Usr>

Returns the USR for the entity that contains the code completion context for this set of code completion results, if applicable.

fn get_results(&self) -> Vec<CompletionResult>

Returns the code completion results in this set of code completion results.