Enum clang::CallingConvention

#[repr(C)]

pub enum CallingConvention {
    Unexposed,
    Cdecl,
    Fastcall,
    Pascal,
    Stdcall,
    Thiscall,
    Vectorcall,
    Swift,
    PreserveMost,
    PreserveAll,
    Aapcs,
    AapcsVfp,
    IntelOcl,
    SysV64,
    Win64,
}

Indicates the calling convention specified for a function type.

Variants
Unexposed

The function type uses a calling convention that is not exposed via this interface.

Cdecl

The function type uses the x86 cdecl calling convention.

Fastcall

The function type uses the x86 fastcall calling convention.

Pascal

The function type uses the x86 pascal calling convention.

Stdcall

The function type uses the x86 stdcall calling convention.

Thiscall

The function type uses the x86 thiscall calling convention.

Vectorcall

The function type uses the x86 vectorcall calling convention.

Only produced by libclang 3.6 and later.

Swift

The function type uses the calling convention for the Swift programming language.

Only produced by libclang 3.9 and later.

PreserveMost

The function type uses a calling convention that perserves most registers.

Only produced by libclang 3.9 and later.

PreserveAll

The function type uses a calling convention that preverses nearly all registers.

Only produced by libclang 3.9 and later.

Aapcs

The function type uses the ARM AACPS calling convention.

AapcsVfp

The function type uses the ARM AACPS-VFP calling convention.

IntelOcl

The function type uses the calling convention for Intel OpenCL built-ins.

SysV64

The function type uses the x64 C calling convention as specified in the System V ABI.

Win64

The function type uses the x64 C calling convention as implemented on Windows.