Enum clang::TypeKind

pub enum TypeKind {
    Unexposed,
    Void,
    Bool,
    CharS,
    CharU,
    SChar,
    UChar,
    WChar,
    Char16,
    Char32,
    Short,
    UShort,
    Int,
    UInt,
    Long,
    ULong,
    LongLong,
    ULongLong,
    Int128,
    UInt128,
    Float,
    Double,
    LongDouble,
    Nullptr,
    Complex,
    Dependent,
    Overload,
    ObjCId,
    ObjCClass,
    ObjCSel,
    ObjCInterface,
    ObjCObjectPointer,
    Pointer,
    BlockPointer,
    MemberPointer,
    LValueReference,
    RValueReference,
    Enum,
    Record,
    Typedef,
    FunctionPrototype,
    FunctionNoPrototype,
    ConstantArray,
    DependentSizedArray,
    IncompleteArray,
    VariableArray,
    Vector,
}

Indicates the categorization of a type.

Variants

Unexposed

A type whose specific kind is not exposed via this interface.

Void

void

Bool

bool (C++) or _Bool (C99)

CharS

The char type when it is signed by default.

CharU

The char type when it is unsigned by default.

SChar

signed char

UChar

unsigned char

WChar

wchar_t

Char16

char16_t

Char32

char32_t

Short

short

UShort

unsigned short

Int

int

UInt

unsigned int

Long

long

ULong

unsigned long

LongLong

long long

ULongLong

unsigned long long

Int128

__int128_t

UInt128

__uint128_t

Float

float

Double

double

LongDouble

long double

Nullptr

nullptr_t (C++11)

Complex

A C99 complex type (e.g., _Complex float).

Dependent

An unknown dependent type.

Overload

The type of an unresolved overload set.

ObjCId

id (Objective-C)

ObjCClass

Class (Objective-C)

ObjCSel

SEL (Objective-C)

ObjCInterface

An Objective-C interface type.

ObjCObjectPointer

An Objective-C pointer to object type.

Pointer

A pointer type.

BlockPointer

A block pointer type (e.g., void (^)(int)).

MemberPointer

A pointer to a record member type.

LValueReference

An l-value reference (e.g. int&).

RValueReference

An r-value reference (e.g. int&&).

Enum

An enum type.

Record

A record type such as a struct or a class.

Typedef

A typedef.

FunctionPrototype

A function prototype with parameter type information (e.g., void foo(int)).

FunctionNoPrototype

A function prototype without parameter type information (e.g., void foo()).

ConstantArray

An array type with a specified size that is an integer constant expression.

DependentSizedArray

An array type with a specified size that is a dependent value.

IncompleteArray

An array type without a specified size.

VariableArray

An array type with a specified size that is not an integer constant expression.

Vector

A GCC generic vector type.