Module clang::sonar

Functions for finding declarations in C translation units.

Structs

Enum
An enum declaration.

Struct
A struct declaration.

Functions

find_enums
Returns the enums in the supplied C translation unit.

find_functions
Returns the functions in the supplied translation unit.

find_structs
Returns the structs in the supplied C translation unit.

find_typedefs
Returns the typedefs in the supplied translation unit.