Struct clang::Type

pub struct Type<'tu> { /* fields omitted */ }

The type of an AST entity.

Methods

impl<'tu> Type<'tu>

pub fn get_kind(&self) -> TypeKind

Returns the kind of this type.

pub fn get_display_name(&self) -> String

Returns the display name of this type.

pub fn get_alignof(&self) -> Result<usize, AlignofError>

Returns the alignment of this type in bytes.

Failures

this type is a dependent type

this type is an incomplete type

pub fn get_offsetof<F: AsRef<str>>(
&self,
field: F
) -> Result<usize, OffsetofError>

Returns the offset of the field with the supplied name in this record type in bits.

Failures

this record type is a dependent type

this record record type is an incomplete type

this record type does not contain a field with the supplied name

pub fn get_sizeof(&self) -> Result<usize, SizeofError>

Returns the size of this type in bytes.

Failures

this type is a dependent type

this type is an incomplete type

this type is a variable size type

pub fn get_argument_types(&self) -> Option<Vec<Type<'tu>>>

Returns the argument types for this function or method type, if applicable.

pub fn get_calling_convention(&self) -> Option<CallingConvention>

Returns the calling convention specified for this function type, if applicable.

pub fn get_canonical_type(&self) -> Type<'tu>

Returns the canonical type for this type.

The canonical type is the underlying type with all "sugar" removed (e.g., typedefs).

pub fn get_class_type(&self) -> Option<Type<'tu>>

Returns the class type for this member pointer type, if applicable.

pub fn get_declaration(&self) -> Option<Entity<'tu>>

Returns the AST entity that declared this type, if any.

pub fn get_element_type(&self) -> Option<Type<'tu>>

Returns the element type for this array, complex, or vector type, if applicable.

pub fn get_fields(&self) -> Option<Vec<Entity<'tu>>>

Returns the fields in this record type, if applicable.

pub fn get_pointee_type(&self) -> Option<Type<'tu>>

Returns the pointee type for this pointer type, if applicable.

pub fn get_ref_qualifier(&self) -> Option<RefQualifier>

Returns the ref qualifier for this C++ function or method type, if applicable.

pub fn get_result_type(&self) -> Option<Type<'tu>>

Returns the result type for this function or method type, if applicable.

pub fn get_size(&self) -> Option<usize>

Returns the size of this constant array or vector type, if applicable.

pub fn get_template_argument_types(&self) -> Option<Vec<Option<Type<'tu>>>>

Returns the template argument types for this template class specialization type, if applicable.

pub fn is_const_qualified(&self) -> bool

Returns whether this type is qualified with const.

pub fn is_elaborated(&self) -> Option<bool>

Returns whether this type is an elaborated type, if it can be determined for certain.

pub fn is_pod(&self) -> bool

Returns whether this type is plain old data (POD).

pub fn is_restrict_qualified(&self) -> bool

Returns whether this type is qualified with restrict.

pub fn is_variadic(&self) -> bool

Returns whether this type is a variadic function type.

pub fn is_volatile_qualified(&self) -> bool

Returns whether this type is qualified with volatile.

pub fn visit_fields<F: FnMut(Entity<'tu>) -> bool>(&self, f: F) -> Option<bool>

Visits the fields in this record type, returning None if this type is not a record type and returning Some(b) otherwise where b indicates whether visitation was ended by the callback returning false.

pub fn is_integer(&self) -> bool

Returns whether this type is an integer type.

pub fn is_signed_integer(&self) -> bool

Returns whether this type is a signed integer type.

pub fn is_unsigned_integer(&self) -> bool

Returns whether this type is an unsigned integer type.